+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [2.1.0] - 2025-08-19
 ### Added
 - Added `BatchableJobQueueInterface` which defines a JobQueue capable of putting jobs to a queue in batches.
   - Extends `JobQueueInterface`.